initial changes for refactoring
